Addendum fix for bug#18352514:
   STRUCT MEMBERS IN MT-SCHEDULER ARE NOT CACHELINE ALIGNED AS ASSUMED

Previous fix failed on Mac OS as there were platform dependent
assumption of the max size of the NdbMutex struct used in
the thr_mutex struct.

This fix introduce the usage of the mysql defined MY_ALIGNED macros
which allows us to more explicit specify the cache line alignment
requirements for the struct members in the scheduler structs.
Some previous padding of these structs has been removed.


****** Patch part 2: ***********
Optional cleanup part of fix for Bug#18352514 	


It used to be defined both a non-aligned 'struct thr_data'
and the aligned 'struct thr_data_aligned'. This was partly
due to historical reasons, and partly because there were 
no simple manners to declare an aligned struct directly -
It has to first be declared as 'normal' struct, then
an aligned variant could be declared from that one by
adding padding.

By using the MY_ALIGNED decl, we could define the aligned
variant directly. The now aligned variant was really only
a helper struct, and has now been removed.