Fix for Bug#19793998 ASSERTION FAILED: ROWS >= 0.0 IN                      HANDLER::INDEX_SCAN_COST()
The assert in Handler::index_scan_cost() fails due to the rows argument
to the function has the value -NaN. The -NaN rows estimate is caused by
a division by zero in the following code in
Optimize_table_order::find_best_ref():

            {                              /* Prefer longer keys */
              cur_fanout=
                ((double) tab->records() / (double) distinct_keys_est *
                 (1.0 +
                  ((double) (table->s->max_key_length-keyinfo->key_length) /
                   (double) table->s->max_key_length)));

The reason for the division by zero is that when creating a temporary table
for the derived table "a" in the query:

  select sha(a) in (select * from (select '' from t1) a) from t1;

the only field in this table has length zero. To be able to use index
access on this table, an index on this field is created. This index will
have length zero and the max_key_length for this table will also be zero.

This fix for this problem is to not create an index on zero-length fields
when adding potential indexes to derived tables.