Bug#19586047: CRASH IN ARG_COMPARATOR::SET_CMP_FUNC
1) A code path in SELECT_LEX::convert_subquery_to_semijoin, which was
only intended for scalar subqueries and row subqueries, was previously
taken by all kinds of subqueries. Add an extra condition to make sure
that it is only taken if the subquery is a scalar subquery or a row
subquery.

2) If an IN predicate whose left side contained a non-scalar row
subquery, had been converted to a semijoin, it would run into an
assert when checking if the semijoin could use materialization. To
avoid the problem, stop trying to use materialization for this kind of
query.