Bug#19766312: SET ENVIRONMENT VARIABLE MYSQL_TZINFO_TO_SQL IN MTR
Problem: There was no refer to the mysql_tzinfo_to_sql binary in test
scripts that would work both when running tests from a sandbox and
from a package. The test was unstable because it used zoneinfo files
from the system where the test was run.

Solution: Set the environment variable MYSQL_TZINFO_TO_SQL to the
correct path to the executable in mysql_test_run.pl in the same way as
is done for other utilities. Use this variable to refer to the binary
in test scripts. Add selected zoneinfo files to std_data/zoneinfo and
modify the test script to use these files.

Create a new test, sys_mysql_tzinfo_to_sql, which uses the system
zoneinfo resources, but which excludes the exact values from the
output so that the test becomes stables across platforms.