Bug#18932813: SIMPLE QUERY AND ITS JSON EXPLAIN GIVES 2 ASSERTIONS Bug#19779600: ASSERT FAILED IN REPLACE_SUBCONDITION WITH IN PREDICATE               INSIDE IN PREDICATE
SELECT_LEX::resolve_subquery uses SELECT_LEX::resolve_place to
determine if an IN predicate is at the AND-top-level of a WHERE clause
or an ON clause, which is one of the requirements for making the IN
predicate a candidate for flattening into a semi-join.

Although resolve_place can be used to determine if the predicate is in
a WHERE clause or an ON clause, it does not reliably tell if the
predicate is at the AND-top-level, since many of the fix_field()
functions don't remember to reset resolve_place before they descend
into their children.

Because of this, some IN predicates were incorrectly flagged as
candidates for semi-join flattening, and that caused assert failures
later when the flattening was attempted.

The fix makes the following changes in order to prevent incorrect
flagging of candidates:

1) Stop using resolve_place to determine if a predicate is at the
   AND-top-level. Only use resolve_place to determine if the predicate
   is in a WHERE clause or an ON clause.

2) Stop switching resolve_place to RESOLVE_NONE for the parts of the
   query tree where semi-join flattening should be disabled. Instead,
   introduce a new field in SELECT_LEX specifically for disabling
   semi-join flattening, called semijoin_disallowed.

3) Switch semijoin_disallowed to true in the cases where resolve_place
   previously was switched to RESOLVE_NONE.

4) Additionally, switch semijoin_disallowed to true when resolving
   parts of the query tree where semi-join flattening previously
   wasn't disabled, but where it should have been disabled because the
   predicates were not on the top-level (such as the left side of an
   IN predicate, arguments of aggregate functions, and arguments of
   LIKE and REGEXP expressions).