Bug#19605472 -  max_statement_time (memory/sql/thd_timer) leaks                 memory
Description:
------------
Issue here is, while cancelling the windows timer the state of the
timer is set to "signalled" in function "my_timer_cancel".
Because of this in "sql_timer.cc" layer, timer is set as
non-reachable and responsibility of freeing memory allocated
for "THD_timer_info" is left to the "timer_callback". But
"timer_callback" function is not called when timer is cancelled.
Hence memory leak is found on Windows platform.

Fix:
------------
Changed code of "my_timer_cancel" function for  Windows so that,
state of the timer is set to "signalled" when timer is expired
(or signalled). Otherwise timer state is set to "non-signalled". 
With this, when timer state is non-signalled, "THD_timer_info"
object is reused. On next "set" operation of timer, new windows
timer is created for it in "my_timer_set" function. Memory
allocated for it is freed in "thd_timer_destroy" operation.
When timer state is "signalled", "THD_timer_info" is freed in
"timer_callback" function. On next "set" operation of timer,
new "THD_timer_info" object and windows timer is created.