Bug #18708334 REMOVE THE SPAM WARNINGS: CLIENT FAILED TO PROVIDE ITS CHARACTER SET
Description:
When we try to connect mysql-5.6 server with mysql-4.0 client the error log gets filled with below warning message
"[Warning] Client failed to provide its character set. 'latin1' will be used as client character set"

Analysis: Since the the client protocol is older than 4.1 and the 4.0 client is not providing the character set while trying to connect to the server. so, the warning message is sent to error log file.

Fix: Since the warning message is not necessary, we are removing the warning message.