Bug #18449085 : WRONG VALUE AFTER MULTI UPDATE. ISSUE : ------- Consider a multi-update statement with 3 tables to join in order t2, t3, t1. And let there be a condition t1.c1 = t2.c2. In the given update statement t2 is the first table and t1 is the third table in join order. We have a condition between them t1.c1 = t2.c2 and let there be a primary key on t1.c1. This condition is pushed down as part of ref access key of t1's look up. Hence t2.c2 is read for every fresh scan of t1. If we update t2.c2 in-between two scans of t1 the condition t1.c1 = t2.c2 will fail as value of t2.c2 has been changed(updated).
So in above case t2 should not be updated immediately after
fetching the rows.
      
SOLUTION:
---------
If a column of first table is part of reference key used by
following tables in join order and also same column is set as
part of update. And then update will be done using a
temporary table.