Bug#19931126: VALGRIND REPORTS USE OF UNINITIALIZED VALUE IN               MY_WILDCMP_BIN_IMPL
The LIKE operator accepts ESCAPE clauses that contain expressions that
are constant at execution time. However, it only evaluates the
expression in the ESCAPE clause if its value is known at resolve time.
If the ESCAPE clause contains an expression that is constant at
execution time, but unknown at resolve time, the escape character will
be uninitialized, and the LIKE operator will produce unreliable
results. This is a regression in 5.6.

The fix factors out the code that initializes the escape character
from Item_func_like::fix_fields() into a new function. If the escape
expression is constant and known at resolve time, it is initialized by
fix_fields() as before. If its value is not known at resolve time,
Item_func_like::val_int() will now initialize the escape character the
first time it is called.

The fix also makes the range optimizer skip the optimization with LIKE
if the escape character is unknown at resolve time. Otherwise, it
would use the uninitialized value for the escape character and produce
unreliable results.