Bug#14571512 : --PRINT_DATA TRUNCATING THE LAST 16 BYTES OF BLOB FIELD
NDB_RESTORE when given with the option --PRINT_DATA, TEXT and BLOB column values
are always truncated to the first 240 bytes instead of 256 bytes.
This is caused due to the consideration of header size (16 bytes) as a part of the
inline size, removing this solves the problem. 
  
The patch includes the fix for this bug along with updated result file of
ndb_restore_conv_more test, which does $NDB_SELECT_ALL (It uses the same code) and
needed to be updated.