Bug#18036143 - CRASH (SEGFAULT) IN ST_JOIN_TABLE::GET_SJ_STRATEGY ON                      UPDATE STATEMENT
Problem:
JOIN_TAB pointer is not checked before de-referencing it.
      
Analysis:
The fact that JOIN_TAB objects are not used for single-table
UPDATE and DELETE, is ignored when optimizer is trying to
generate simple equality condition from multiple
equalities in eliminate_item_equal().
As a result server fails accessing the uninitialized pointer.