WL#7640  Ndb Active Active Delete-Delete handling
Implementation and testcase enhancement for extra status variable

  ndb_conflict_last_stable_epoch

This is added to give guidance to the user in the unexpected 
primary failure scenario.

The Secondary cluster's last_stable_epoch is the last Secondary
epoch which was committed, binlogged, replicated and reflected 
without any conflicts being noted.

Assuming such an epoch exists, knowing which it is it may help 
in handling any alignment issues during an uncontrolled 
Secondary promotion to Primary.

ndb_rpl_conflict_epoch2_extra enhanced to cover expected
behaviour.