Fix Club test regressions after backport of bug#19524096:
Paste from a mail describing the issue below.
As it is yet not clear wheteher this fix should also
go into 7.3+ (as discussed below), it will for now
only be pushed to 7.1 and 7.2 as a regression fix.

....................................

I have analyzed this and found the root cause to be the
'lock()' set in ClusterMgr::is_cluster_completely_unavailable()

....
bool
ClusterMgr::is_cluster_completely_unavailable()
{
  bool ret_code = true;
  lock();  <=

....


In most circumstances 'is_cluster_completely_unavailable' is
called while holding the poll right. As part of your (Mikael R)
ATC patches (WL#3860) the implementation of the 'PollGuard'
changed:

Prior to WL#3860 the (global) TransporterFacade::theMutexPtr is held
when owning the poll right. This WL changed that such that only
the per-client mutex trp_client::m_mutex is locked.

The above 'lock' in ::is_cluster_completely_unavailable() will call
trp_client::lock(), which in 7.1 is implemented as:

void
trp_client::lock()
{
  NdbMutex_Lock(m_facade->theMutexPtr); <- Already held
  ...
}

Thus, we will try to re-lock TransporterFacade::theMutexPtr
which we already held, and will either block forever, or
fail due to 'lock already held' (Os dependent?)

In 7.3+, the 'lock' call will lock ClusterMgr::trp_client::m_mutex
instead, which will be granted **if the current client is not ClusterMgr itself**.
So this might not be entirely safe either, depending on the which
trp_client calling checking is_cluster_completely_unavailable().
AFAIK, no test failures observed yet though.

Furthermore, looking at the implementation of ::is_cluster_completely_unavailable()
I cant really find any reason for this 'lock' to be needed at all.
What we need to protect is the 'NodeInfo' maintained in ClusterMgr::theNodes[].
Thus, locking clusterMgrThreadMutex should be sufficient - both
in order to fix the redness introduced in the 7.1-> backport,
and as an fix/improvement to 7.3.