Bug#19903481 - INTERNAL ERROR DURING ONLINE REORG, TRIX (LINE: 782) 0X00000002
Fix bug in bug fix.

Some SUMA scans can return TRANSID_AI with zero-length attribute(s).

This results in the accumulated data section having length 0.

appendToSection() with length zero does not allocate a section.

However the code in sendScanSubTableData() assumes that there is always
a section.

Further, the recipients of SUB_TABLE_DATA expect that there will always
be two sections, and the 'Data' section may have length zero. (TRIX, API)

A good fix would be to allow the Data section to be missing.

A quick fix is to send a 'zero length' section in this case.

Fixes failures in
  ndb.ndb_fk_cascade_delete_multi
  ndb.ndb_fk_deferred_cascade

Push to 7.1+