WL#7592 step 4. GTIDs: Generate Gtid_log_event and Previous_gtids_log_event always. Refactor @@session.gtid_executed implementation.
==== Background ====

Before this worklog, every transaction cache or statment cache is
associated with a Group_cache. The Group_cache contains the GTID of the
transaction. (It is actually implemented to hold a sequence of GTIDs,
which was a preparation in case NDB would use that, but we have not
finalized the design of how NDB integrates with GTIDs yet, so currently
it is only a list of one element.) The *session* variable
@@session.gtid_executed contains the GTIDs that are currently in the
Group_cache.

An equivalent characterization of @@session.gtid_executed is: it is the
value of @@session.gtid_owned, in case the thread owns any GTID and the
transaction cache or statement cache is non-empty.In all other cases,
@@session.gtid_executed is an empty string.

In a later step of this worklog we will make Gtid_log_event be generated
always. To do this, we have to hold the Gtid_log_event outside of the
transaction cache or statement cache. This in turn means that the
Group_cache will not hold the GTID of the transaction. So it would break
the current implementation of @@session.gtid_executed.

==== Patch ====

In the present patch, we change the implementation of
@@session.gtid_executed so that it does not read from the Group_cache.
Instead, it just checks if the statement cache or transaction cache is
nonempty; if one of them is, it reads @@session.gtid_next. Otherwise, it
returns the empty string.

Thus, this patch prevents the patch that generates Gtid_log_event always
from breaking @@session.gtid_executed.

In addition, Sys_var_gtid_executed currently inherits from
Sys_var_gtid_set_func. It does not need to do that with the new
implementation, so we make it inherit from Sys_var_charptr_func instead.

Now, Sys_var_gtid_set_func becomes unused, so we remove it.

@sql/binlog.cc
- We don't need the function to get the Group_cache any more, since
  the function was only used by Sys_var_gtid_executed.
- But we need a new function, is_binlog_cache_empty, which checks if
  both the statement cache and the transaction cache are empty.

@sql/sql_class.h
- Update prototype and description.

@sql/sys_vars.h
- Remove unneeded Sys_var_gtid_set_func.
- Make Sys_var_gtid_executed check the state of the
  statement/transaction cache directly, and read the GTID from
  @@session.gtid_owned instead.