WL#7592 step 6. GTIDs: Generate Gtid_log_event and Previous_gtids_log_event always. Allow writing Log_event header to memory.
This is a small refactoring to allow writing Gtid_log_event to a memory
buffer rather than to an IO_CACHE. This does not change the behavior of
the server, so it is not possible to have a test case for this.

When, in a later patch, we write Gtid_log_event always, the
Gtid_log_event will live outside the transaction cache. Before the event
is flushed to the binary log, it has to be stored in a memory buffer.
However, there was no function for writing log events to a memory
buffer. This patch adds the function Log_event::write_header_to_memory
that writes the 19 byte Log_event header to a memory buffer, and
Gtid_log_event::write_to_memory that writes an entire Gtid_log_event to
a memory buffer.

@sql/log_event.cc
- Move code for writing the Log_event header out of Log_event::write and
  into the new function Log_event::write_header_to_memory.
- Move body of Gtid_log_event::write_data_header into
  new function Gtid_log_event::write_data_header_to_memory.
- Simplify the logic for clearing flags&~LOG_EVENT_BINLOG_IN_USE_F
  before generating the checksum.

@sql/log_event.h
- Add Log_event::write_header_to_memory.
- Add Gtid_log_event::write_header_to_memory.
- Add Gtid_log_event::write_to_memory.
- Make Log_event::write_header, Log_event::write_footer,
  Log_event::need_checksum, and Gtid_log_event::print private,
  since there is no reason for them to be public.