WL#7592 step 7. GTIDs: Generate Gtid_log_event and Previous_gtids_log_event always. Refactor MYSQL_BINLOG::do_write_cache.
This patch rewrites do_write_cache so that it becomes more maintainable
and so that we can write out-of-cache events to the binary log. This
does not change server behavior, so it is not possible to have a test
case for this.

Rationale 1:
Before this patch, do_write_cache mixes two tasks:
 1. It reads from the statement or transaction cache and assembles
    pieces of events that are split over multiple pages in the
    IO_CACHE.
 2. It processes the event and writes it to the binary log.

In WL#7592, we need to store the Gtid_log_event outside the statement
and transaction caches. The event needs the processing in (2), but not
that in (1). Hence, we de-couple these two tasks. Task (2) is
implemented by a new class and task (1) remains in do_write_cache.

Rationale 2:

Before this patch, do_write_cache was very difficult to understand, as
it kept lots of state in variables that were updated throughout the
function. The reason was that it was organized in an unpractical way:
the outermost loop iterated over pages of the IO_CACHE and tried to keep
various pieces of state of half-written events in local variables. This
state information was updated all over the function, which made it
difficult to understand what were the loop invariants.

This patch changes so that the outermost loop iterates over events. This
makes the state information is much more short-lived. For instance,
state related to the beginning of an event is only needed at the
beginning of the iteration and can be stored in a very short-lived
variable. So there is less state kept between iterations, and this makes
the loop invariants much simpler.

@sql/binlog.cc
- Add auxiliary class Binlog_event_writer, which handles fixing the
  fields before writing to the binlog.
- Remove fix_log_event_crc, which was only doing trivial things and
  is now part of Binlog_event_writer.
- Add auxiliary function read_cache_page used by do_write_cache.
- Change do_write_cache to use Binlog_event_writer and be more
  maintainable.
- Declare the Binlog_event_writer a couple of levels up in the call
  stack, in binlog_cache_data::flush, and pass it as a parameter to
  write_cache and do_write_cache. The Binlog_event_writer is needed
  in binlog_cache_data::flush in the next patch.
- Move DBUG_EXECUTE_IF("simulate_binlog_flush_error", ...) out of
  write_cache, into the caller. This is important since this is
  supposed to be executed prior to writing anything in the transaction.
  Before the patch, write_cache was the only function that wrote the
  transaction. However, after next patch, Gtid_log_event will live
  outside the cache and will be written to the binary log before we call
  write_cache, so then the error simulation must happen before writing
  the Gtid_log_event.

@sql/binlog.h
- Fix whitespace in declaration of wait_for_update_bin_log.
- Make do_write_cache private since there is no reason for it to be
  public.
- Add Binlog_event_writer parameter to do_write_cache and write_cache.