Fix regression in AutoTest -n DropWithTakeover T1 introduced by fix for bug#19874809:
The above test started failing after pushing fix for 19874809 :
'DICT operations during 'takeover' may crash new master'

That fix changed how we handled a SCHEMA_TRANS_END_REQ on 
a new master still on the process of handling MASTER_TAKEOVER.

 - Previously a REF('Busy') were sent during takeover.
   The API client handled that REF as a Schema transaction failure
   instead of waiting for the final rollforward/backward in the
   takeover process. Thus, loosing the knowledge of the transaction
   outcome.

 - With the fix we now wait for the takeover to complete, and then
   sends a SCHEMA_TRANS_END_REP reporting either a commit or abort.

The problem here is that the testcase 'DropWithTakeover' delays
the DROP_TAB_REQ indefinitely. This breaks the requirement that 
every REQ should be either CONFed or REFed, or a NODE_FAILURE
received.
Furthermore, the testcase incorrectly required that 'dropTable()'
failed. This used to be due to a REF(Busy) was interpretted as 
transaction failure. However, in this case the dropTable is rolled
forward by the new master, and actually succeeds.

Fix changes the delay of DROP_TAB_REQ from being indefinite to
a 10s delay which should be sufficient for the master to crash.