Bug#19955501 - INNODB_ZIP.WL6501: ASSERTION FAILED: VECTOR SUBSCRIPT OUT OF RANGE
The class truncate_t contains a vector of indexes called m_indexes. Each
index_t in that vector contains a vector of bytes descibing each field.  This
vector is type fields_t and is called m_field. m_field is set initially to
NULL and then fields are added to it in truncate_t::parse(), but only if the
table is compressed.

Then in truncate_t::create_indexes(), the address of the first field in
m_fields is referenced like &m_fields[0] to initialize a btr_create_t object
called btr_redo_create_info.  But if this is not a compressed table, the
number of elements in this vector is zero and the [0] subscript is indeed out
of range.

This error is reported by Windows and should be reported in Linux.  Either
way, NULL is returned, which is what the btr_create_t object expects if there
are no fields to define for this index.

The solution is to just return NULL when the table is not compressed rather
than &it->m_fields[0].

Approved by marko in rb#7290