Bug #19880569: CIRCULAR INCLUDE DEPENDENCY BETWEEN SQL_LEX.H AND SQL_CLASS.H
This is an internal refactoring.

Problem: The header files sql_lex.h is was not self-sufficient, and required
sql_class.h to be included before it, always. Meanwhile, sql_class.h was
including sql_lex.h in the middle of the file. The reason is a design flaw:
sql_class.h was defining enums that are needed by the parser and InnoDB,
both of which should be insulated from sql_class.h. The enums are

- enum Key::Keytype
- enum enum_ha_read_modes
- enum enum_filetype
- enum Foreign_key::fk_match_opt
- enum Foreign_key::fk_option

Fixed by moving the enums to sql_lex.h and removing all inclusions of
sql_class.h where possible; it was obviously included just to be able to
include sql_lex.h. We also make sql_class.h include sql_lex.h at the top, to
make the dependency one-way.