Bug#19803418   ASSERTION: TABLE->QUIESCE == QUIESCE_NONE              IN ROW0QUIESCE.CC LINE 683
PROBLEM

When flushing the table,innodb tries to store all locks
on the table.If we have multiple read locks on the table
(which we can achieve using aliases),we try to set the
quiesec state of the table to QUIESCE_START multiple times
which causes the assert.

FIX

Relax the assert if table is already in QUIESCE_START
state.

[Approved by Jimmy rb#7167]