Bug #19528825 "UNABLE TO PURGE A RECORD" WITH IBUF AND VARCHAR FIELDS CONTAINING SPACES
Problem:

This problem happens only if the change buffer is enabled.  It does not
happen if the change buffer is disabled.  Analysis showed that the cause
was the buffering of the following sequence of operations for a same
record - IBUF_OP_DELETE_MARK, IBUF_OP_INSERT, IBUF_OP_DELETE.  Because
of this sequence, while merging the change buffer entries into the page,
the delete operation (also called purge) found the "record to be purged"
without a delete mark.

Solution:

The above sequence of operations must not be bufferred.  This sequence is
avoided when change buffer is disabled, because of latching of secondary
index page.  So it must be avoided when change buffer is enabled by using
the "page watch" properly.

Once the purge thread has kept a watch for a page in the buffer pool,
a concurrent insert operation on the same page must take this into account,
and not buffer the operation.

rb#7293 approved by Marko.