Bug#19912474 - MYSQL_EMBEDDED FAILS WHEN LINKER SELECTS WRONG VERSION OF HANDLER CONSTRUCTOR
Before this fix, class handler contained members defined conditionally
with #ifdef HAVE_PSI_TABLE_INTERFACE.

Even when the members defined like this are private,
changing the binary layout of the class causes build issues
when linking code compiled with/without the performance schema
instrumentation.

This was found while building the embedded server,
but the same issue also exists when compiling a server
and storage engine plugins with different options.

The fix is to always define the attributes needed
by the performance_schema table io instrumentation,
even when these attributes are not used.

A consequence is that the structure PSI_table_locker_state
is now always defined in psi.h