WL#7083 step 3.6. GTID_MODE settable: Reduce locking in save_gtids_of_last_binlog_into_table.
Before this patch, save_gtids_of_last_binlog_into_table held
global_sid_lock.wrlock while updating a table. This is too long to hold
a lock, code-level locks should not be held while updating a table as
that could lead to deadlocks. The only reason the lock was held was to
look up sidno in global_sid_map.

Fixed this by using a local sid_map instead. This does not require a
lock, so we can release the lock earlier.

@rpl_gtid_state.cc
- Use a local sid_map instead of the global one.

@rpl_gtid_set.cc
- Enable existing but disabled code for using Gtid_set::add_gtid_set and Gtid_set::remove_gtid_set with two sets having different Sid_maps.