Bug#19191377: WL#6737: NOT NULL COLUMN IS BEING PROJECTED AS               NULLABLE IN TEMPORARY TABLES Introduced more details on such a bug fix:
There are two error places which cause this bug.
1)From the issued query, we can see 't2.f1' can be NULL because the query is
a LEFT JOIN.  However due to the Not-NULL-where condition (t2.f1 > 1), the query
will be converted into an inner-join. After transformation, 't2.f1' can't be NULL
any more because t2.f1 has a NOT-NULL constraint. But in mysql, this field is
still treated as might-be-null. This is incorrect. This is similar with an old
bug#53793.

2)There is an code error for keyinfo->key_length calculation after wl6711,
which doesn't calculate NULL bytes at all. NULL bit is considered when all
fields are created. But before the bug fix, keyinfo->key_length has finished its
calculation before <hash_field> is created. This is incorrect.

For the above two reasons, we can see 2) must be fixed. After 2) is fixed,
NULL bytes will be calculated. Even 1) still exists, it can't introduce a crash
any more. Fixing 1) need a refactoring work on might-be-null attribute of field
after query rewrite. It needs a big work. So I choose to fix 2) at first. After
fix bug#53793, 2) will be resolved simultaneously.